Check for taximeter being initialized before ending travel, also changed the order or where the initialization is checked on the start trip as it first checked wether it was with status for hire instead of initialized first which would easily cause confusion to anyone using the app

+++ app/src/main/java/com/driverapp/fragments/ShiftsFragment.kt
@@ -145,21 +145,30 @@ class ShiftsFragment : Fragment(),
             })
     }
 
-    override fun onDisplayExtendedStatus(p0: Any?, displayExtendedStatusResponse: DisplayExtendedStatusResponse?) {
+    override fun onDisplayExtendedStatus(
+        p0: Any?,
+        displayExtendedStatusResponse: DisplayExtendedStatusResponse?
+    ) {
         exStat = displayExtendedStatusResponse?.extendedStatusData
-//        shiftNo.text = "TURNI: " + (exStat?.ShiftNumber ?: "N/A").toString()
     }
 
     override fun onFullShiftDetailsResponse(p0: Any?, response: FullShiftDetailsResponse?) {
-        val totalTripsCount = (response?.fullShiftDetails?.ShiftInfoEnd?.TripsQuantity?.toInt() ?: 0) - (response?.fullShiftDetails?.ShiftInfoStart?.TripsQuantity?.toInt() ?: 0)
-        val totalFareAmount = (response?.fullShiftDetails?.ShiftInfoEnd?.TotalAmount?. toDouble() ?: 0.0) - (response?.fullShiftDetails?.ShiftInfoStart?.TotalAmount?.toDouble() ?: 0.0)
-        shiftNo.setText("TURNI: " + (response?.fullShiftDetails?.ShiftConsecutiveNumber ?: "N/A").toString())
+        val totalTripsCount = (response?.fullShiftDetails?.ShiftInfoEnd?.TripsQuantity?.toInt()
+            ?: 0) - (response?.fullShiftDetails?.ShiftInfoStart?.TripsQuantity?.toInt() ?: 0)
+        val totalFareAmount = (response?.fullShiftDetails?.ShiftInfoEnd?.TotalAmount?.toDouble()
+            ?: 0.0) - (response?.fullShiftDetails?.ShiftInfoStart?.TotalAmount?.toDouble() ?: 0.0)
+        shiftNo.setText(
+            "TURNI: " + (response?.fullShiftDetails?.ShiftConsecutiveNumber ?: "N/A").toString()
+        )
         tripCount.setText("UDHETIMET: " + totalTripsCount)
         totalFare.setText("TOTALI: " + totalFareAmount)
         Log.d("ShiftsFragment", "Full Shift Details: ${response?.fullShiftDetails?.ShiftInfoStart}")
     }
 
-    override fun onLastClosedShiftDetailsResponse(p0: Any?, lastClosedResponse: LastClosedShiftDetailsResponse?) {
+    override fun onLastClosedShiftDetailsResponse(
+        p0: Any?,
+        lastClosedResponse: LastClosedShiftDetailsResponse?
+    ) {
         //Log.d("LastClosed", "Shift Details: ${lastClosedResponse?.fullShiftDetails.toString()}")
         //
         //val request = OldTripShiftRequest(
